+++ aviary/subsystems/aerodynamics/gasp_based/test/test_gasp_aero_coeffs.py
@@ -3,8 +3,12 @@ import unittest
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_check_partials, assert_near_equal
 
-from aviary.subsystems.aerodynamics.gasp_based.gasp_aero_coeffs import AeroFormfactors
-from aviary.variable_info.variables import Aircraft, Mission
+from aviary.subsystems.aerodynamics.gasp_based.gasp_aero_coeffs import (
+    AeroFormfactors,
+    FormFactor,
+    BWBFormFactor,
+)
+from aviary.variable_info.variables import Aircraft
 
 tol = 1e-8
 
@@ -38,6 +42,56 @@ class TestAeroCoeffs(unittest.TestCase):
 
         partial_data = prob.check_partials(method='cs', out_stream=None)
         assert_check_partials(partial_data, atol=1e-12, rtol=1e-14)
+
+
+class FormFactorTest(unittest.TestCase):
+    """Test fuselage form factor computation and SIWB computation"""
+
+    def test_case1(self):
+        prob = om.Problem()
+
+        prob.model.add_subsystem(
+            'formfactor',
+            FormFactor(),
+            promotes=['*'],
+        )
+
+        prob.model.set_input_defaults(Aircraft.Fuselage.AVG_DIAMETER, val=19.365, units='ft')
+        prob.model.set_input_defaults(Aircraft.Fuselage.LENGTH, val=71.5245514, units='ft')
+
+        prob.setup(check=False, force_alloc_complex=True)
+        prob.run_model()
+
+        tol = 1e-5
+        assert_near_equal(prob[Aircraft.Fuselage.FORM_FACTOR], 1.35024726, tol)
+
+        partial_data = prob.check_partials(out_stream=None, method='cs')
+        assert_check_partials(partial_data, atol=1e-11, rtol=1e-11)
+
+
+class BWBFormFactorTest(unittest.TestCase):
+    """Test fuselage form factor computation and SIWB computation"""
+
+    def test_case1(self):
+        prob = om.Problem()
+
+        prob.model.add_subsystem(
+            'formfactor',
+            BWBFormFactor(),
+            promotes=['*'],
+        )
+
+        prob.model.set_input_defaults(Aircraft.Fuselage.HYDRAULIC_DIAMETER, val=19.365, units='ft')
+        prob.model.set_input_defaults(Aircraft.Fuselage.LENGTH, val=71.5245514, units='ft')
+
+        prob.setup(check=False, force_alloc_complex=True)
+        prob.run_model()
+
+        tol = 1e-5
+        assert_near_equal(prob[Aircraft.Fuselage.FORM_FACTOR], 1.35024726, tol)
+
+        partial_data = prob.check_partials(out_stream=None, method='cs')
+        assert_check_partials(partial_data, atol=1e-11, rtol=1e-11)
 
 
 if __name__ == '__main__':
